Add icon User list and flight list selection throughout testig Replace the toast messages with progress bar Add the Google Maps showing Handle JSON parse exceptions Introduced POJO data objects for the Android, as the annotated ones resulted in a wild compiler error due to a compiler bug Changed to boxed objects all the fields of the data objects to allow the introduction of non-existing fields

+++ fs_proxy/src/fs_proxy/FS_Proxy_Basic.java
@@ -175,11 +175,11 @@ public class FS_Proxy_Basic implements FS_Proxy_I {
                 .simulationTimeStamp(simulationTimeStamp)
                 .aircraftTypeName((String) readFromGeneral(FSInformation.AIRCRAFT_TYPE_NAME))
                 .aircraftCode((String) readFromGeneral(FSInformation.AIRCRAFT_CODE))
-                .altitude(aircraft.Altitude())
+                .altitude((double) aircraft.Altitude())
                 .heading(p)
                 .longitude(aircraft.Longitude())
                 .latitude(aircraft.Latitude())
-                .verticalSpeed(aircraft.VerticalSpeed())
+                .verticalSpeed((double)aircraft.VerticalSpeed())
                 .build();
     }
